Change return type of GetAvarageByAge to int

+++ ConsoleApp8/ConsoleApp8/DoctorService.cs
@@ -68,9 +68,9 @@ namespace ConsoleApp8
             return doctorArray;
         }
 
-        public double GetAvarageByAge(Doctor[] doctorArray)
+        public int GetAvarageByAge(Doctor[] doctorArray)
         {
-            double result = 0;
+            int result = 0;
             foreach(var doctor in doctorArray)
             {
                 result += doctor.Age;
